OSL implementation of RGB ramp node.
The sampled color ramp data is passed to OSL as a color array. This has to be done as actual float[3] array though, since the Cycles float3 type actually contains 4 floats, leading to shifting color components in the array.

Additional parameter set functions for arrays have been added to the Cycles OSL interface for this purpose.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@50637 954f8c5b-7b00-dc11-b283-0030488c597c