Mark movie clip user current frame as editable
It should be pretty safe change which would allow doing stuff like
python-defined tracking routines without need to update the whole
scene when it's needed to perform some operation on different clip
frame.

It'll be possible to write operators similar to tracking, which
updates space clip's frame number, but not scene frame when tracking
and only synchronizes scene frame number on operator finish.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@50749 954f8c5b-7b00-dc11-b283-0030488c597c