Fix #32742: Motion path calculation on linked armatures locks up Blender
Issue was happening when linking armature object and making proxy and was
caused by not copying visualization settings in BKE_pose_copy_data.

This lead to deadlocks in motion path drawing code.

After discussion with Campbell decided it is crucial fix since it fixes
bug appearing in really common scenario of using armatures.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@51015 954f8c5b-7b00-dc11-b283-0030488c597c