Fix/workaround #32846, dupli group + particle instances gets messed up in Cycles viewport rendering.
Caused by modifier updates during dupli-list generation. The dupli-list generation temporarily changes the ob->obmat matrix, which in turn leads to wrong particle states if used for reset. Skip the particle update if no timestep is performed or initialization required.

Proper solution for this problem would be to avoid changing the object data (= particles) state altogether in modifiers, which are usually only writing to DM data and not touching the object or base mesh. This would require a well designed physics framework and integrating it into current particles is close to impossible.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@51368 954f8c5b-7b00-dc11-b283-0030488c597c