Camera: bring Title Safe display up to date, making it a Safe Areas option that shows both title safe and action safe areas following more modern standards.
Patch #32822 by Harley Acheson, full description:

Our current "title safe" camera display option is anachronistic. It shows a
border of 10% on all edges, which used to be the recommended title safe area
for 4:3 content on standard definition CRT televisions. However we are very
unlikely to create new projects that output for SD TV at that aspect ratio.

This patch change the option to "safe areas" with and indicates the
"title safe" area (also known as "graphic safe") as well as the "action safe"
area. "Title Safe" is an area visible by all reasonably maintained sets, where
text was certain not to be cut off. "Action Safe" is a larger area that
represented where a "perfect" set (with high precision to allow less
overscanning) would cut the image off.

The current recommendation for Action Safe is 3.5% on all edges, which is the
maxiumum overscan for TVs now. The recommended title safe is now 5% vertically
and 10% horizontally for content that is of wider aspect ratio than 4:3. The
reason for the difference between horizontal versus vertical margin is because
wider content would be letterboxed on an older 4:3 television, giving it
additional margin.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@51623 954f8c5b-7b00-dc11-b283-0030488c597c