Fix #33217, mirror modifier edges missing in wireframe mode.
I see we explicitly set the edge visibility in the code for a few
modifiers because of this flag missing. The only place that this flag is
not set is during subsurf so maybe it would make a lot more sense to set
it by default on newly created edges (currently off by default) through
bmesh and turn it off for any modifiers that may need it to be off.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@52399 954f8c5b-7b00-dc11-b283-0030488c597c