Fix crash opening some .blend files on OS X 10.8 with double click or drag and drop onto the application.
It seems something changed in the operating which makes our method of releasing
windows crash. Previously we called [m_window release], but on 10.8 this does
not remove the window from [NSApp orderedWindows] and perhaps other places,
leading to crashes. So instead we set setReleasedWhenClosed back to YES right
before closing, which will then do the cleanup for us.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@52781 954f8c5b-7b00-dc11-b283-0030488c597c