Bugfix #33541 - Deleting all keyframes leaves dangling action groups
When deleting all keyframes in F-Curves, the corresponding F-Curves are deleted.
If all the F-Curves in an action group were deleted in such a way, the group
wouldn't be removed. This meant that these groups would never be shown (until
F-Curves for these groups were created again), but would still exist, causing
problems when trying to rearrange groups in the animation editors (i.e. groups
would appear to not move). Now these groups get deleted when they get empty.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@53052 954f8c5b-7b00-dc11-b283-0030488c597c