Bugfix #33458
Tooltips in Blender were not scaling when zooming in/out on UI elements.
(They did follow DPI though, but tooltips were in a 'global' fixed size).

Error was that for tooltips on large scaled popups (like in Node editor)
the drawing was entirely wrong even.

Now tooltips scale correctly, also for fonts.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@53060 954f8c5b-7b00-dc11-b283-0030488c597c