fix/workaround [#33493] checker de-select and edge loops
Checker de-select by default would give uneven selection on a circle, this isnt really a bug but the offset used would give unevenly spaced selection even if the 3rd vertex for eg could be evenly selected on a circle.

Change how the offset works so the active element always remains selected when the offset is set to zero, this tends to give more even de-selection.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@53086 954f8c5b-7b00-dc11-b283-0030488c597c