From the long todo:
New Outliner mode: "Main Data".

This shows a flattened, non-hierarchical list of all linkable "ID" data in
your current project. It works fine on searches. Actually this is the
view on the "Main" database in Blender, the one that's saved in a .blend.

This is in general more useful than the "Datablocks" viewer,
which is not searchable, and shows every property of data as well.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@53266 954f8c5b-7b00-dc11-b283-0030488c597c