Pack UI:
Oops! So there was an Unpack button, hidden in the Nkey properties of Image window.
This was drawn next to greyed-out buttons, didn't notice it well.

I also now found the unpack() menu in editors/util, and there's an Image unpack op.
Also the RNA api has an unpack!

Will remove the generic unpack op, and add a sound and vfont unpack instead.




git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@54123 954f8c5b-7b00-dc11-b283-0030488c597c