Fix #34041: rendered view + border
Issue was caused by rare cases when camera move happens just after
last sample was finished, this would lead to missing delay reset
because render cycle will go to pause_cond.wait(). No reset will
happen at this point because of some kind of optimization which
checks whether camera is tagged for update and wouldn't do reset
in this case.

Talked to Brecht and seems this optimization is not actually needed
and removing it will solve issue with frozen preview.

git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@54326 954f8c5b-7b00-dc11-b283-0030488c597c