Bugfix [#34123] Armature "Switch Direction" doesn't work when selected bone belongs to more than one chain
For example:
               /----->C
   A-->B-:
               \----->D

If bone B is selected, then it would get operated on twice, creating the
illusion that it had not been operated on. This is because we traverse up the
chains (child to parent) as the EditBone structure only stores parent to
children relationships only. A second invocation of this operator would then
work fine, as all the links to other bones would have been removed, thus
preventing further problems.

Fixed by tagging bones that have been operated on.


git-svn-id: https://svn.blender.org/svnroot/bf-blender/trunk/blender@54345 954f8c5b-7b00-dc11-b283-0030488c597c